request #16234: API explorer should be the default `api` link
As of today when one click on "API" link in footer, navbar or help
access they are directed to an almost empty page where there is 2
sentences, one that refers to tuleap doc and one with a link to API
explorer.

Most of the time people are looking for API explorer then it should be
the default link.

In the explorer, at the beginning we should have the same "API help"
intro with reference to Tuleap documentation.

The old SOAP documentation links are moved under /soap/ url.

If "api_explorer" plugin is not installed, there are no links to REST
documentation.

Change-Id: Ia3c178f504c4e13038a212294b854ee9d6722b1a