Display a spinner when loading the test execution in the TTM UI
Currently there is nothing except a blank rectangle which gives the
impression something is broken. The issue is particularly visible when
accessing directly to a test execution in a relatively large campaign.

This situation can for example happen when using the "Got to the last
execution" link in test plan.

Took the opportunity to rework a bit the loaders of test management to
make sure the image used as a loader can be properly cached by browsers.

Part of story #15094: have more effective test plan

Change-Id: Ie85a3fafb7c1fb1c678e8b387e7dc266c6c45eb5