Push story and activity in  "➕ New" dropdown
Run forgeupgrade, existing "story" trackers are put in
the "➕ New" dropdown (this is not the case for existing
"activity" trackers, users will have to do this by hand).

---

Create a new project based on AgileALM template.
You should be able to add a new story across the project via the "➕ New"
dropdown.

Same for Kanban template, with activity tracker.

Part of story #16205: « + New » button

Change-Id: I148c30b99a3a2f562ef229fcaeeea253881c3416