request #16567: IE11 cannot display default avatars
The default avatars are generated in JPEG but the headers announce a PNG
image preventing IE11 to decode the image.

This contribution makes sure the default avatar are generated in PNG.
A new command is also introduced to remove all the default avatars. It
can be used by instances to fix existing avatars and would also us to
propagate changes to the default avatars in the future.

Change-Id: I4334c9155e77d29e3cfdf7107254cca7b8dd6802