Use system font instead of Source Sans Pro
Part of request #16597 Use system font and rem unit

This commit remove the usage of Source Sans Pro font
so we don't have to load an extra font to display Tuleap pages.

It fixes font sizes everywhere in TLP. Plugins and
Core pages will be done in a next commit.

To improve the accessibility in Tuleap, font sizes, line height…
are now defined in rem or em so the base font size defined in
the browser settings is took into account.

Change-Id: Ib2cbc58f50171e8cbde67ae593c5a2510fa4cfae