request #16136 Workflow post actions not triggered at artifact creation when workflow field is not submittable
In a User Story tracker configure state "On going" with "add to top backlog"
post action.

In field configuration, define "On going" as default value of "status"
and remove the "can submit" permission of this field.

Create a new user story in tracker

* It has the "On going" status
* but the artifact does not appear in top backlog

This issue only occurs if the field used in workflow is not submitable.

Change-Id: Idb60dd1a344fd662449ad6090a5247d669aa4d42