Introduce the artifact link type _mirrored_milestone
This type of link is going to be used to track milestones that are a
mirror of a milestone in a aggregator project.

This type of link is not supposed to be manually set/added by a user nor
be disabled by a project admin. Users can however remove an art link of this
type when they edit an artifact.

To test you will need to set this type at least once in an artifact. To
do that you either need to play directly with the
tracker_changeset_value_artifactlink table or with the request sent when
editing/creating the artifact because the backend does not enforce the
capability for a user to add/edit specific art link type 🤷.

Part of story #16125: [release management] create a new Aggregator Release

Change-Id: I63904254bedf26b41d620fc9dfafb20755964743