Prevent desactivation of the artifact link types at the project level
The multi-project backlog is going to use artifact links to manage the
mirrored milestones and allowing the desactivation of the types of link
will be a source of trouble.

Once a project has enabled the artifact link types, it is no more
possible to go back.

Part of story #16125: [release management] create a new Aggregator Release

Change-Id: Ieacdcf331b37a1ae9c5a92e0f183ba69d4f40b47