Prevent usage of a non editable artifact link type on the backend
As of today even if the artifact link types are not displayed in the UI
they can still be used. The backend and frontend must be consistent in
the behavior.
With this contribution, it's no more possible to submit or edit a link
to use an non editable type unless the link is already using it.

The easiest way to test it is to edit the DOM to change the value of the
selected type for something non editable (with the multi-project backlog
enabled you can use "_mirrored_milestone").

Part of story #16125: [release management] create a new Aggregator Release

Change-Id: I5bef15c51cc72cf58a903034c2e9e519b8675a05