Display project flags in breadcrumbs
This is necessary to be able to remove project flags from navbar

If sidebar is collapsed, project flags are displayed before the
breadcrumb.

To not clutter the review, the removal from the navbar and the display
of project flags in the expanded sidebar will be done in a dedicated
contribution.

Note: color of the project flags/privacy will be done later.

Part of story #16210: Remove navbar

Change-Id: I61ceaeb28b5216b4b9dda14d27a58bb08633d14f