request #16612 CVS commits are no more stored in centos 7
If you have your Tuleap on centos 7 and still using CVS, we strongly advise to no more use it.
But if you really want / need to continue to use it, then the commits are no more stored on DB.
We need to use sudo to have the script running well.

To test this patch, you have to:

* Having a working CVS service on your c7 platform
* Deploy src/utils/sudoers.d/tuleap_cvs_commit_prep and
  src/utils/sudoers.d/tuleap_cvs_log_accum into /etc/sudoers.d/
* Redeploy commit_prep and log_accum into /usr/lib/tuleap/bin/
* Run the 2 following commands:
  - find /cvsroot/*/CVSROOT -name "loginfo" -exec sed -i "s#ALL /usr/lib/tuleap/bin/log_accum#ALL sudo -u codendiadm -E /usr/lib/tuleap/bin/log_accum#" {} \;
  - find /cvsroot/*/CVSROOT -name "loginfo" -exec sed -r -i "s#ALL \(/usr/lib/tuleap/bin/log_accum (.*)\)#ALL sudo -u codendiadm -E /usr/lib/tuleap/bin/log_accum \1#" {} \;
* Enable commit storage on your CVS repository
* Do a commit
* The commit must be stored with the right username

Change-Id: I74260468caed5da2f83b38a8a016c85825d673e4