request #16605: Planning v2, kanban and Test Management cannot be used with IE11
Also a lot of dates are just not visible at all when using IE11 or Edge
Legacy (e.g. in the document or tracker plugin).

Dependending on the situation either the element is loaded without the
custom elements polyfill (that is the case for the crash of the Test
Management plugin) or only the polyfill is loaded (that is the case
whenever the date is not visible).

It seems that approach that was took when introducing a custom element
for the first time and reused with the relative-date element is
depending too much on developers being vigilant and that does not scale.

This contribution changes the approach that was taken: instead of
relying on developers to determine when the polyfill should be loaded,
we now detect if the feature is present and if it is not the polyfill is
loaded asynchronously. The impact is minimal for modern browsers as they
only need to load and execute the feature detection code before loading
the rest of the component. For deprecated browsers it is a bit less nice
because it means they will need to wait for the polyfill to be lazy
loaded before continuing (currently they directly have access to it
without needing to emit another request). They have a way to solve that:
stop using a deprecated browser.
For the developers the only time it becomes necessary to remember a
polyfill might be needed is when creating the custom element instead
of every time the custom element is used.

Planning V2 is affected by another issue independent from this one so it
still does not work.

Change-Id: I919bd0797539366c839688eb9c167e917def0aed