Introduce Switch to modal
This is a base commit to build the Switch to modal.

Modal should look nearly the same between BurningParrot and
FlamingParrot.

The modal only displays "Under construction" therefore it is hidden
temporarily behind the lab mode. This restriction will be removed in the
next commit.

Modal can be open with:
* Switch to button in the navbar
- `/` or `s` shortcuts. If one is taken by your browser, try the another
one − e.g `/` is taken to search in the page on my Firefox.

Note: empty presenter is intended, it will encapsulate usefull data later.

Part of story #16209: « Switch to… » button

Change-Id: Ic985e542f906435cb430a0be8fbaad2229de7bca