Tuleap Core provides a mean to identify the authentication mean
The main goal it to be able to refer to LDAP as LDAP in OpenID Connect
Client login screen when used in combination with LDAP auth.

This takes the opportunity to introduce yet another variable in LDAP
config to customize the name of the LDAP server in deployment context.

Part of request #16607 Allow ldap login when openidconnect client is used

Change-Id: I67fdf363d3cfbad7a260e74c70a50c8238d54843