Adjust log levels and contexts in the MilestoneCreatorChecker class
The info level messages have been moved to the debug level as they can
happen quite often on a busy instance and are not that useful outside of
a debugging session.

Logs about a user missing missing permissions to do an action has also
been moved to the debug level as it is not the result of a incorrect
configuration and can perfectly happen in the day-to-day usage.

Exceptions are given as part of the context so that the logs contains
the full information.

Part of story #16125: [release management] create a new Aggregator Release

Change-Id: I2cd12aba8cfa1a94376497180ff62a3deb790020