Display list of projects in switch to dropdown
If user does not have any projects => empty state is displayed.
Else display the list of projects according to mockup (dedicated design
check is planned anyway).

If user can access the software map, a link is displayed.

Filter will come in a dedicated changeset.

No moar Lab mode to have the switch to.

Part of story #16209: « Switch to… » button

Change-Id: I3ccfa0f68ca73d7a4c718268db8f774765f8dfbc