request #16617: Tuleap service can crash if it is started before the DB
While this is to be expected this is annoying when booting the server
and the DB server is hosted on the same server.

Note that we cannot set a dependency to the DB service because it might
be hosted on a different server. It is also not possible to use the
systemd feature 'Restart=on-failure' because the systemd version used on
CentOS/RHEL7 is too old and does not support it.

When testing, do not forget to tell systemd to reload the configuration
with 'systemctl daemon-reload'.

Change-Id: Ia2bbd05f7f618d34817d069f4229bb926cbf2f96