Add a global error state in planning view
part of story #16125 [release management] create a new Aggregator
Release

How to test:
- rebuild the planning v2 app & scss
- When there is an issue while loading the initial milestones, a global
error state will be shown and will replace the rest of the app. No more
infinite spinner.
- The error state shows the RestException error message if any and lets
people try to reload the page

Note to reviewer: loading closed milestones does not have the same error
state. It should probably be an "error modal" like in other apps because
"open" milestones and left-hand side backlog items are already loaded.
The error modal does not exist yet in planning v2. It may be introduced
later in order to ease this review.

Change-Id: I0cb353ac25aabda233dec4dda529aa1bb892c6df