LDAP + OIDC accounts are created like LDAP accounts
When an LDAP account is created there are a number of things that are done
behind the scenes to guarranty the correctness in all cases (esp. svn).

This commit introduce changes in the way accounts are created by ldap plugin
so when OIDC Client plugin create accounts, the result is compatible (esp.
regarding `ldap_id` and `plugin_ldap_user` tables).

This assume that OIDC provider returns an attribute in the `userinfo` route
that can be used as `ldap_id`. If such an attribute exists it can be
configured with

    tuleap config-set openidconnectclient_ldap_attribute family_name

Some events were converted to objects (instead raw string) to ease the reasonning.
The modification of `account_redirect_after_login` (add PFUser as param and pass it
in the event was mandatory because when OIDC plugin called this method, at the other
side of the event, in ldapPlugin, UserManager::getCurrentUser() was not up to date.

Part of request #16607 Allow ldap login when openidconnect client is used

Change-Id: I9170d0b3362b8062f4bb4e8ac66b6ffe8b743de8