request #16623 Docman download as zip must keep the uploaded filename
When you upload a file into the document service of your project,
you can add a custom name.
If this name does not contain the file extension, then the "download as zip"
feature will add this file with its custom name without any extension.

However, when downloading by hand the file, we use the FS name and not the custom one.

Both features must behave the same, so the uploaded file will be used in the archive.

Change-Id: I9d4032fb9a0e733040b783e0e77b06933bb10709