Do not highlight the release note the first time a user login with a OIDC provider
This is a follow-up to 5485d5343f58d6ea6dec821fc0d96ad95647a1d5.

In a situation where an instance as set an OIDC provider as the only
authentication method the release note should not be highlighted the
first time the user sign-in.

Part of story #15102: highlight informations to the end user

Change-Id: I108c11ea96e5aa0f7fa110b25c7b7ef118200741