Display recent items in "Switch to…" modal
To test:
* visit a user story
* visit a kanban

Open the "Switch to…" modal: user story is displayed with a xref, kanban
is displayed with an icon.

Only basic styling has been done, but it should be "usable" though.
* color of the icon/xref will come in a dedicated changeset
* filter will come in a dedicated changeset
* skeletons styling will come in a dedicated changeset
* overflow when title of the item is very long will come in a dedicated
changeset

Note: in order to avoid repeating myself with style definition in BP and
FP, I created a common scss file. Only stuff related to the current
changeset has been put into it. A dedicated changeset will move the
remaining common definitions.

Part of story #16209: « Switch to… » button

Change-Id: Icf45ad1b894d05d3fc143340dc8d2bc04237eda2