Cannot use a synchronized field in a workflow transition rule in contributor tracker
When the situation happens, it should not be allowed to create a new
milestone in the aggregator project.

For now only the workflow transition rules are verified, other workflow
behaviors such as field dependencies or global rules will be added
ulteriorly.

Part of story #16124: [release management] ensure Aggregator release creation is possible

Change-Id: I8d5ddde310594c0e2a8c66988927507a2a57aaa6