Refactor BeforeLogin event
Object based events are easier to deal with.

This refactoring introduce a behaviour change. For plugins that take over
the authentication (eg. ldap), there is no longer a difference between
'login exists but password is wrong' and 'login do not exists'. In practice
this shouldn't change anything as the interface doesn't make any difference
and it's a bad security practice to distinguish the 2 cases anyway.

Part of request #16607 Allow ldap login when openidconnect client is used

Change-Id: Iec6b177b2348240c7a8d946535e2c525f76e0e56