Forbid Local or LDAP login when a user authenticate with OIDC
As long as OIDC is used Tuleap should not have to deal with user
password. Hence if a local or LDAP authentication is attempted
it should be rejected in favor of:
- OIDC flow for web login
- Access Keys for Git & REST

Please note that, at this stage, svn still authenticate against
the database or LDAP.

How to test: Once your account is linked to an OIDC provider you
should no longer be able to use your previous credentials (local
db or LDAP) for git over http, rest calls or web UI login. SVN
login with those credentials should still work though.

Please note this is per user. Users that nevered link their accounts
to OIDC should authenticate as usual.

Part of request #16607 Allow ldap login when openidconnect client is used

Change-Id: I9667955ea050de8ac04cda04ec8a45285bc1004a