Fix frs-plugin cypress fail
request #16606 Correct errors seen in nightly tests

make test_cypress
=> check result in <workspace>/test_results_e2e_full

About the fix:
It's crap, but I don't know how to do best:
when we submit a request is thrown in ajax, thats what we catch with the @createRelease route
after the form is submitted, I don't understand why, but the redirection is done before the release is created in db, so we are redirected on an empty package.

Wisiting the full service, fix the issue
I didn't do antyhing funky here, I just check the differnecies between code exectued in frs plugin and in frs.spec.ts and I use the same trick.

Change-Id: I19a3a6814fd51c89f2046c94006db40316546523