request #16642 Bump slugify from 1.3.6 to 1.4.5
No functional changes

Changelog:
- https://github.com/simov/slugify/releases/tag/v1.4.0
- https://github.com/simov/slugify/releases/tag/v1.4.1
- https://github.com/simov/slugify/releases/tag/v1.4.2
- https://github.com/simov/slugify/releases/tag/v1.4.3
- https://github.com/simov/slugify/releases/tag/v1.4.4
- https://github.com/simov/slugify/releases/tag/v1.4.5

I'm not sure to understand why, but in previous version call with '-' were properly replaced in tests.
It does no longer works with the bump, it seems that we should use the replacement character '_' and not '-' in order to have our test suite works.
(BTW in both cases it works properly on UI).

Change-Id: I1932bf7f89838550ec0f5bfd59ed97140cda0c5c