Allow plugins to inform about login being refused
When LDAP or OpenID Connect refuse a login attempt, they should be able
to give an hint to end user that might get nuts trying to authenticate.

Remove dead code in ldap afterLocalLogin. The code was no longer reachable
because it was relying on ldap_id but the first condition already asserted
it was empty.

Part of request #16607 Allow ldap login when openidconnect client is used

Change-Id: I92f320121034f9306e7c02f810352ecc35bd384e