Document should not rely on navbar-height
Try to drag n drop a file in the current folder in Document. The
highlight should be glued to the top (or just under the project navbar,
if any).

The remaining usage of $navbar-height is kept to better identify where
we need to adjust margin. In that case the quicklook that has a fixed
position when we scroll is too close to the floating button. A dedicated
patch will handle that.

Part of story #16210 Remove navigation bar

Change-Id: I55ae02cfba73eb6fe9636d551d57e2a5f62bab29