Kanban should not rely on navbar-height
Furthermore when sidebar is collapsed, the $breadcrumb-height is not
accurate. We now have a dedicated variable in that case. It is defined
in px, but should probably be defined in rem to follow font zoom. It
should be handled in a dedicated commit though.

Part of story #16210 Remove navigation bar

Change-Id: I48825bc93910ec460e24c010768282c4087b5bad