Limit usage of $navbar-height in taskboard
The height of the app does not anymore depends on navbar which is now
floating.

However we must keep the $navbar-height usage for fixed swimlane
computation since that we need to have a pinned header so that floating
buttons do not overlap the swimlane headers on scroll. The pinned header
is the same that homepage one for example but with taskboard
specificities.

Part of story #16210 Remove navigation bar

Change-Id: I8d073228b62dd134b6ceb861fd2bc317c17ee31a