Remove toolbar from artifact view
Most of the links are already present in the dropdown menu in the
breadcrumbs (and creation link is also present in the New dropdown).

Note: "Contextual" help link is removed to be consistent with latest feature
introduced in Tuleap.

The remaining information in the toolbar was the create by mail popover.
This is now moved in the dropdown menu and opens a modal. In order to be
consistent, the reply by mail popover becomes also a modal (see
follow-up comments textarea in an artifact).

Since the breadcrumb is also available on burning parrot pages (for
example tracker » admin » workflow » transition) the modal should also
be displayed here.

Part of story #16210 Remove navigation bar

Change-Id: Ia6d232faa8ef20a9986fd02e441877e4e2f978ef