Default logo is in SVG format
If site administrator didn't customize the default logo, then we should
deliver a SVG version so that:
* there is less chance for blur
* we can customize the color according to the current theme

A logo is considered customized as soon as the organization_logo.png is
present in /var/lib/tuleap/images/ AND different than the one delivered
by default in BurningParrot theme.

The abilitiy to customize the svg logo (for example to let the site
administrator use the company logo in svg) will come in a dedicated
changeset.

For now the detection of customized logo is done for each request. While
the size of the files are small, it is inefficient to do it for each
request for each user. A cache will be put in place but in a different
commit.

Part of story #16210 Remove navigation bar

Change-Id: Id6113e89a4ce6ee0d05ea7f1e58221bc92e8aa53