request #16688 In the artefacts history, only the first field change can be detailed
When you have 2 changes in text fields in the same changeset,
only the first one can be displayed in the follow-up.

In addition, I removed the ability to open field changes when accessing
comment URL directly. This worked only if we had one text field to open,
but not if we have multiple. It can still be reintroduced if necessary.

Change-Id: I95c2e907268786a6608ffe22019787d15953fb78