Remove mediawiki toolbar
Now administrators must use the breadcrumb dropdown menu.

Without this patch, due to absence of the toolbar, regular users see the
floating buttons on top of mediawiki menu. Now with the breadcrumb that
is displayed for everybody, the issue is gone.

(It is not ideal though since there is a global issue with breadcrumb
and floating buttons alignment, but it is at least consistent with the
other pages.)

Part of story #16210: Remove navigation bar

Change-Id: I49257d3ffd024f064f307a576cc47e98b720cb1d