request #17055: Creating a campaign should not pollute your history
Go to TTM service of a project where there is a lot of tests.

Create a new campaign with "all tests".

Submit.

Refresh the page and open the SwitchTo modal: only the new campaign
artifact should appear at the top of recent items. New test exec
artifacts should not appear; those are technical items and don't need to
appear there.

The same when you add or create a new test in an existing campaign.

Change-Id: I49e57e6938a73d2459f61ddab5e2b9e323e97ce4