request #17078: Too much lookalike breadcrumbs in git
Given a namespaced repository (e.g deps/tuleap/rhel/6/php-mediawiki-tuleap)
When you fork the repository
And you go on the fork,
Then since Tuleap 12.0 there is too much lookalike breadcrumbs.

We can safely remove the repository path since it is already displayed
in the breadcrumbs. The "Forked from" information is kept as is until we
know if the current proposal is good enough in the fields.

Change-Id: I29a08843d509c8ebf13245a7d388fb3db26411f4