Fix request #17087: big motd + logged in + resize = 💩
Replace jScrollPane by smooth-scrollbar in FP. That way we use the same
library than BurningParrot. One less jQuery dependency, yay \o/

Fix issue when you have a very long motd text, and you are logged in, and you
resize the viewport: the logo should not go under the motd when the
latter wraps.

Change-Id: Ic5f4812cb4214664c19c48238fbbeee9193c731e