request #17089: Remove (for real) timer input from the test execution display
In c4866f6b22ec20007fe277cf784f17fe1d8d331e this timer was supposedly
removed but the commit message was lying the timer was just hidden.
This contribution takes care of removing the code.

Enabling and disabling the "saving" status of a test execution has been relocated
in the controller to be sure the change back to the non saving mode is
correctly detected even where there is no change on the execution
itself. It also makes the mechanism easier to reason about as the switch
between the saving and non-saving modes is local instead of being
splitted between the controller and the service.

Change-Id: I1a042cf4570cf0365f24a0f589dadb8d6ed04bb1