Add back type annotation on TestsDataFromJunitExtractor::extractTestExecutionData
The type annotation was removed in
e5026b250be2b54fe210a9e23efd68c21dade211. Widening a type is not the
right solution in this situation as it reduces the possibilities of
catching issues with static analysis.

Part of request #17067: PATCH on /testmanagement_campaigns/{id} should take test suite into account

Change-Id: I8d919b0d7db78aa9a10df779c29dc90b93b7f437