Split milestone tracker collections
part of story #16125 [release management] create a new Aggregator
Release

No functional change expected. The "+ Add Releases" button only appears
if all checks are successful.

This is a refactoring to have a "Contributor Tracker Milestone
collection" that clearly represents what it contains. It will be useful
when creating the "mirror releases" as we don't want to include the
aggregator release in the mix.

Change-Id: Iba07290055cfe02476bffad06a97a6771299af12