request #17101: Avatar in current user menu shouldn't be shrinked
If your real name is long, then the avatar may appear oval instead of
round.

This is not the first time the issue occurs. Every time we fixed the
situation with a `flex: 0 0 auto;`. Instead of repeating ourselves and
omitting this rule the next time we will display an avatar, the rule has
been moved to tlp-avatar* base component.

Took the opportunity to ellipsis the realname if it is really too long.

Change-Id: I3b8bd5a3b84c4b18c3a75a80065ff052b038070a