request #17104: Cannot access to Git tag information
Issue introduced by git #tuleap/stable/0497e3de63860e3e2aab754d65f66153baa677d5.

The possibility to view the content of a tag is re-introduced with the
same look and feel than the rest of the repository viewer.

URLs to reach the tagged object is kept as is, but may lead to errors. It
will be handled later.

Change-Id: I707fc41b8dde3a45efd9b23bf2d31cc4e23adcea