We don't need mustache to render a list-picker
part of: story #16645 have searchable (multi) select boxes with (static binding, no dependencies)

Mustache was creating more problems than it solved. Now it's gone, here
are the benefits:
- calls to Element.querySelector has been drastically reduced
- no need to check that the result of my selection is instanceof Element
  anymore
- we create the base components of the list-picker once, then pass them
  to the classes that need it -> complexity reduced
- A small refactoring was possible, so i moved the creation of the base
  components in a dedicated class.
- This class can be now used to write
  unit tests of classes and functions manipulating these elements.

I also made sure that the base components of the list picker have the
same name in each class, so we know which element we are manipulating.

And last, but not least, the dropdown width is not assigned manually
anymore, it is all done using CSS.

-- No functional change expected

Change-Id: I0209dca2de05235edd9c1f2ce89bdca7e64e78ab