request #17107: Export features of a tracker should require an account
The export features of the content of a tracker are quite resource
intensive and a few abusive users can affect negatively the experience
of all the other users.

While this contribution does not solve the core issue (a specific set of
feature should not have the possibility to affect this much a whole
instance) it makes easier to manage abusive users on public instances.

Change-Id: Ia02ce2ac8e3aebe28acdc4cecb5096e49499fe8c