request #17113: Workers complain they cannot write their lock file after a reboot
Since systemd has took over, /var/run is now a tpmfs filesystem so it is
not preserved between reboots. If parts of the app expect to write in a
folder like /var/run/tuleap, the folder needs to exist first. systemd can
take care of that. For reference a similar issue was recently fixed for
Tuleap CVS deployment see git #tuleap/stable/29fe5c7e2862d9602ecaa1a4c3d70587cd234735.

This contribution fixes the issue for the Tuleap workers trying to write
in /var/run/tuleap but also potentially for the ForumML plugin (even if
this one seems fine to not have the folder when it starts its process).

Change-Id: I03168f28b5740fe415a042eabb14405091f93f10