Modal uses the REST route
So that users can actually invite people.

Click on [Invite] button.
Enter some emails in the modal.
Click [send invitation].
=> emails are sent, a success feedback is displayed and you can close
the modal and start over

If some emails are not sent, then a warning message is displayed.

If every emails are not sent, then a general error message is displayed.

FlamingParrot will be done in a dedicated commit.

Part of story #16633: invite buddies to my Tuleap

Change-Id: Ifb73eb3ab9d8ac8a5b526dac5aed3485cb12aaa4