request #17117: Do not allow to link a special account to an OpenID Connect provider
For most of the special accounts it should never even happen since it
should not be possible to login with them so there is no reason to allow
it. However since git #tuleap/stable/2e365854e78066cfaf597785d78eab54e07a18fa
it possible to lock yourself out of an instance if you use the dynamic
credentials plugin. Indeed, if you map the dynamic credential user to an
OpenID Connect provider it will not possible to connect to the account
again at the end of the lease.

Change-Id: I330dfe99026a8863366360f82091239200b618bd