Have a selection state
part of: story #16645 have searchable (multi) select boxes with (static binding, no dependencies)

We need a selection state to ease the manipulation of list items.
To do so, ListPickerItem type now reference the element inside the
dropdown list and the target <option> inside the source <select>

It will avoid us having to query the DOM to know which element is
currently selected, and which <option> we have to select/unselect.

-- no functional change expected

Change-Id: Ie33d049fb5ef5e82ca3a3b2f18d906f4413c4556