Refactoring: configure ArtifactLink validation with context
Part of story #16125 [release management] create a new Aggregator
Release

No functional change expected.

This refactoring has two goals:
1. Break artifact link validation out of the FormElement_Field
inheritance. This allows us to NOT have to modify every field in
Tuleap because we are changing a method signature like validate().
This is partially successful here: ArtifactLinkValidator
is called directly by the general "FieldsValidator". We must still
call $field->validateFieldWithPermissionsAndRequiredStatus() to check
basic things like permissions and "required".

2. Configure artifact link validation with an object. The object
allows us to change the validation (disable some checks) based on
the kind of action we are doing. In the next patch, it will let us
allow the creation of an artifact link that is usually not editable.
Hence the names "SystemAction" where the systemis allowed to create
those links, but for "ManualAction" it is not allowed.

Change-Id: I2bd1c19eaf92b2479985a217e7b8960890f762a7