List site level OAuth2 apps in the site administration
This is a quick first step listing the site level apps in the site
administration. The main goal of the contribution is to make easier to
see which parts are common between the project admin and the site
administration so some elements like the presenters can be reused.

To test you can set the column project_id of the
plugin_oauth2_server_app table to null for one of the existing app.

Part of story #17106: have OAuth2 apps at Site Level

Change-Id: I40b528dbc8de0fdf7b742ee89097cd877756ea61