Convert icons that are subject to BC to FA5 — fixed width edition
With FontAwesome 5, comes a list of breaking change in icon names.
Please see Upgrading from Version 4 documentation for details[0].

With the help of a script[1], we can easily convert the old names to the
new ones.

In order to limit the size of the review, this change targets only:
* fixed width usage of icon (fa-fw)
* direct usage of "fa fa-fw fa-xxxx".
  => Concatenation of strings, complex mustache conditions, js injections, … will be done later.
* icons that are subject to BC
  => remaining icons will be done later

No functional changes.

Part of request #17110: Update to FontAwesome 5

[0] https://fontawesome.com/how-to-use/on-the-web/setup/upgrading-from-version-4#name-changes
[1] https://github.com/estebandelaf/scripts/blob/master/fontawesome4to5.sh

Change-Id: I03518d7c58203151fe34f7c2c3b46d03a1b39a05