Invitation is complete when the user registers
As a user User1, invite an external person to register to Tuleap.
When user User2 register using the email used for the invitation, then
User1 receive a notification.

The same if it is site administrator that create the account for User2.

If User1 send many invitations to the person, only one notification is
sent.

If many users invited the same person, each users receive a
notification.

Part of story #16633: invite buddies to my Tuleap

Change-Id: I6fd908139ed8091be732fd1396a22f5eaeba63a6