Mirror milestone also copies description
Part of story #16125 [release management] create a new Aggregator
Release

How to test:
- Set up multi projects
- In the Aggregator project, create a new Release and fill its
Description field.
- Mirror releases should be created in all the Contributor projects.
Each mirror release should have the same Description filled with the
same content.

/!\ Image upload in the description is not supported (it won't be
copied).

Note to reviewer: this contribution is also mixed with a refactoring to
DRY the usage of semantics in multi project backlog. Sorry for the size
of the review but I don't have the heart to extract it into its own
review yet...
The main goal was to stop copy/pasting everywhere the semantic factories
getting semantics and checking if they aren't null and getting their
fields and checking the fields aren't null.
There should be no functional change in the "Mirror milestone creation
checks".

Change-Id: I7e9b8b8d54ef6673223b25d05fef6acc76606898